requirements.txt
astro-sdk-python==1.8.0
databricks-cli==0.18.0
apache-airflow-providers-databricks==6.2.0

dbt-core==1.7.10
dbt-postgres==1.7.10
astronomer-cosmos==1.3.2
soda-core-spark
soda-core-spark-df